Add files via upload
Some code to work off of, as well as what should've already been uploaded (Earlier phases, wireframe, et cetera.). As of right now I've only stamped out our main panel and how it connects to the other panels - for now just the list of stocks held. This was more just to test how I can navigate between "pages" with the setContentView() method as it's not something we've covered in labs.

This upload is on the 18th of march. I've got to finish off lab 6 for EECS but afterwards I'll be back focusing this. I'll try to have it's skeleton (that is, the app without it's main challenge; it's database, specifically getting information in) done tonight to show in tomorrow's meeting.

For those who're able to work on this; these are some tangible steps you can do atm:

- Detail the stocklist.xml file so as it matches the wireframe, without the stock information - of course.
- Add new panels and detail them similarly for the post-search window. Connect them as is shown in mainactivity.java's goToList() method.
- Work on the aesthetics of any panel, I don't know much in this department (outlines, stylized text, etc.).
- If you've studied java before, create either a text file or xml file that will serve as our fake "database". As well as an ArrayList that holds information parsed from said file. 
  For now just use manually-inserted and arbitrary stock information.
- Anything else you can think of, see the definitions in our phase2 document.